+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next version
 + Change API documentation URL
++ Add custom 404 Exception
 
 ## 0.7.0
 + Add stats route